chore: fix outliers and export missing types
