+++ src/components/shared/useElementEvent.ts
@@ -1,4 +1,4 @@
-import type { Mask } from '../../BaseElementTypes';
+import type { CreateEvent, Mask } from '../../BaseElementTypes';
 import { ElementType } from '../../BaseElementTypes';
 import { useElementValidation } from './useElementValidation';
 import { has, isEmpty } from 'ramda';
@@ -6,40 +6,6 @@ import { useMemo } from 'react';
 import { useCardMetadata } from './useCardMetadata';
 import { extractDigits, isNilOrEmpty } from '../../utils/shared';
 import { _elementErrors } from '../../ElementValues';
-
-type FieldError = {
-  targetId: string;
-  type: 'incomplete' | 'invalid';
-};
-
-type ElementEvent = {
-  /**
-   * `true` if the element is empty.
-   */
-  empty: boolean;
-  /**
-   * `true` if the input `valid` and `maskSatisfied` properties are `true`.
-   */
-  complete: boolean;
-  /**
-   * `true` if the input is valid according to validation for each element.
-   * Defaults to `true` if no validation is defined for the element.
-   */
-  valid?: boolean;
-  /**
-   * `true` if the input satisfies the mask length requirements.
-   * Defaults to `true` if no mask is provided.
-   */
-  maskSatisfied?: boolean;
-  /**
-   * Array of objects that indicates if an element is invalid or incomplete.
-   */
-  errors?: FieldError[];
-};
-
-type EventConsumer = (event: ElementEvent) => void;
-
-type CreateEvent = (value: string) => ElementEvent;
 
 type UseElementEventProps = {
   type: ElementType;
@@ -115,5 +81,3 @@ export const useElementEvent = ({
     };
   };
 };
-
-export type { CreateEvent, EventConsumer, ElementEvent };
